requirements.txt
Django==5.0.6
gunicorn==22.0.0
pytest-django==4.8.0